Refactored DataTable's columns and filter logic
